+++ cliquet/tests/testapp/views.py
@@ -20,3 +20,13 @@ class Toadstool(resource.ProtectedResource):
 class Moisture(resource.ProtectedResource):
     # Empty schema.
     pass
+
+
+class PsilocybinSchema(resource.ResourceSchema):
+    # Optional fields.
+    edible = colander.SchemaNode(colander.Boolean(), missing=True)
+
+
+@resource.register()
+class Psilo(resource.ProtectedResource):
+    mapping = PsilocybinSchema()