+++ cliquet/tests/resource/test_views.py
@@ -259,6 +259,40 @@ class RecordAuthzDeniedTest(AuthzAuthnTest):
                           headers=self.headers, status=403)
 
 
+class EmptySchemaTest(BaseWebTest):
+    collection_url = '/moistures'
+
+    def test_accept_empty_body(self):
+        resp = self.app.post(self.collection_url,
+                             headers=self.headers)
+        self.assertIn('id', resp.json['data'])
+        resp = self.app.put(self.get_item_url(uuid.uuid4()),
+                            headers=self.headers)
+        self.assertIn('id', resp.json['data'])
+
+    def test_data_can_be_specified(self):
+        resp = self.app.post_json(self.collection_url,
+                                  {'data': {}},
+                                  headers=self.headers)
+        self.assertIn('id', resp.json['data'])
+
+    def test_data_fields_are_ignored(self):
+        resp = self.app.post_json(self.collection_url,
+                                  {'data': {'icq': '9427392'}},
+                                  headers=self.headers)
+        self.assertNotIn('icq', resp.json['data'])
+
+
+class OptionalSchemaTest(EmptySchemaTest):
+    collection_url = '/psilos'
+
+    def test_known_fields_are_saved(self):
+        resp = self.app.post_json(self.collection_url,
+                                  {'data': {'edible': False}},
+                                  headers=self.headers)
+        self.assertIn('edible', resp.json['data'])
+
+
 class InvalidRecordTest(BaseWebTest):
     def setUp(self):
         super(InvalidRecordTest, self).setUp()